refactor(#234): split dns.rs into dig/nslookup submodules Split the 1035-line dns.rs monolith into a proper module directory, following the docker/ and gh/ patterns already in infra/. New structure: dns/mod.rs — shared re-exports and test helpers dns/dig.rs — CONFIG_DIG, RE_DIG_*, run_dig(), parse chain, tests dns/nslookup.rs — CONFIG_NSLOOKUP, RE_NSL_*, run_nslookup(), parse chain, tests No cross-calls exist between dig and nslookup; the only shared elements are the load_fixture/make_output test helpers (moved to dns/mod.rs under #[cfg(test)] pub(super) mod test_helpers). infra/mod.rs unchanged — dns::run_dig and dns::run_nslookup resolve via re-exports in dns/mod.rs. All 29 DNS unit tests pass without modification. Co-Authored-By: Claude <noreply@anthropic.com>
